move ToKebabCase to DragonFruit
