+++ src/System.CommandLine.DragonFruit.Tests/StringExtensionTests.cs
@@ -1,11 +1,10 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.CommandLine.Parsing;
 using FluentAssertions;
 using Xunit;
 
-namespace System.CommandLine.Tests
+namespace System.CommandLine.DragonFruit.Tests
 {
     public class StringExtensionTests
     {
