+++ src/System.CommandLine.DragonFruit/StringExtensions.cs
@@ -1,0 +1,63 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Text;
+
+namespace System.CommandLine.DragonFruit
+{
+    public static class StringExtensions
+    {
+        public static string ToKebabCase(this string value)
+        {
+            if (string.IsNullOrEmpty(value))
+            {
+                return value;
+            }
+
+            var sb = new StringBuilder();
+            int i = 0;
+            bool addDash = false;
+
+            // handles beginning of string, breaks on first letter or digit. addDash might be better named "canAddDash"
+            for (; i < value.Length; i++)
+            {
+                char ch = value[i];
+                if (char.IsLetterOrDigit(ch))
+                {
+                    addDash = !char.IsUpper(ch);
+                    sb.Append(char.ToLowerInvariant(ch));
+                    i++;
+                    break;
+                }
+            }
+
+            // reusing i, start at the same place
+            for (; i < value.Length; i++)
+            {
+                char ch = value[i];
+                if (char.IsUpper(ch))
+                {
+                    if (addDash)
+                    {
+                        addDash = false;
+                        sb.Append('-');
+                    }
+
+                    sb.Append(char.ToLowerInvariant(ch));
+                }
+                else if (char.IsLetterOrDigit(ch))
+                {
+                    addDash = true;
+                    sb.Append(ch);
+                }
+                else //this coverts all non letter/digits to dash - specifically periods and underscores. Is this needed?
+                {
+                    addDash = false;
+                    sb.Append('-');
+                }
+            }
+
+            return sb.ToString();
+        }
+    }
+}
