+++ src/System.CommandLine/Parsing/StringExtensions.cs
@@ -9,7 +9,7 @@ using System.Linq;
 
 namespace System.CommandLine.Parsing
 {
-    public static class StringExtensions
+    internal static class StringExtensions
     {
         private static readonly string[] _optionPrefixStrings = { "--", "-", "/" };
         private static readonly char[] _argumentDelimiters = { ':', '=' };
@@ -500,59 +500,6 @@ namespace System.CommandLine.Parsing
             first = null;
             rest = null;
             return false;
-        }
-
-        public static string ToKebabCase(this string value)
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                return value;
-            }
-
-            var sb = StringBuilderPool.Default.Rent();
-            int i = 0;
-            bool addDash = false;
-
-            // handles beginning of string, breaks on first letter or digit. addDash might be better named "canAddDash"
-            for (; i < value.Length; i++)
-            {
-                char ch = value[i];
-                if (char.IsLetterOrDigit(ch))
-                {
-                    addDash = !char.IsUpper(ch);
-                    sb.Append(char.ToLowerInvariant(ch));
-                    i++;
-                    break;
-                }
-            }
-
-            // reusing i, start at the same place
-            for (; i < value.Length; i++)
-            {
-                char ch = value[i];
-                if (char.IsUpper(ch))
-                {
-                    if (addDash)
-                    {
-                        addDash = false;
-                        sb.Append('-');
-                    }
-
-                    sb.Append(char.ToLowerInvariant(ch));
-                }
-                else if (char.IsLetterOrDigit(ch))
-                {
-                    addDash = true;
-                    sb.Append(ch);
-                }
-                else //this coverts all non letter/digits to dash - specifically periods and underscores. Is this needed?
-                {
-                    addDash = false;
-                    sb.Append('-');
-                }
-            }
-
-            return StringBuilderPool.Default.GetStringAndReturn(sb);
         }
 
         private static IEnumerable<string> ExpandResponseFile(
